Fixed dtype in compute_output_spec function to match the dtype of the input

+++ keras_core/ops/nn.py
@@ -1107,7 +1107,7 @@ class MultiHot(Operation):
         )
 
     def compute_output_spec(self, inputs):
-        return KerasTensor(inputs.shape + (self.num_tokens,), dtype=self.dtype)
+        return KerasTensor(inputs.shape + (self.num_tokens,), dtype=inputs.dtype)
 
 
 @keras_core_export(
